requirements.txt
fastapi
uvicorn
pandas
google-generativeai
requests